pyproject.toml
[project]
name = "resume-helper"
version = "1.2.0"
description = "Resume management skills suite: extract, optimize, format, review, and create custom templates"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4",
    "jinja2",
    "pdfplumber",
    "pydantic>=2.0",
    "python-docx",
    "pyyaml",
    "requests",
]

[project.scripts]
resume = "resume_cli.__main__:main"

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/resume_cli"]
